# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v2/tours/{tourId}": {
      "patch": {
        "operationId": "ToursController_update",
        "summary": "Update tour by id",
        "parameters": [
          {
            "name": "tourId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Required idempotency key for update mutation.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTourDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Tours"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ToursController_getById",
        "summary": "Get tour by id",
        "parameters": [
          {
            "name": "tourId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Tours"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v2/settings/equipment": {
      "get": {
        "operationId": "SettingsEquipmentController_list",
        "summary": "List workspace equipment items",
        "description": "Readable by all workspace members (for tour UI). Create/update/delete remain owner/admin only.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceEquipmentItemResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Settings — Equipment"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "SettingsEquipmentController_create",
        "summary": "Create workspace equipment item",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEquipmentItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceEquipmentItemResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Settings — Equipment"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v2/settings/tour-themes": {
      "get": {
        "operationId": "SettingsTourThemesController_list",
        "summary": "List workspace tour themes",
        "description": "Readable by all workspace members (for tour UI). Create/update/delete remain owner/admin only.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceTourThemeResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Settings — Tour themes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "SettingsTourThemesController_create",
        "summary": "Create workspace tour theme",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceTourThemeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceTourThemeResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Settings — Tour themes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "API v2 Documentation",
    "description": "",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v2"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CostContextDto": {
        "type": "object",
        "properties": {
          "currency": {
            "type": "string",
            "example": "USD",
            "description": "ISO 4217 alpha code when set."
          },
          "totalCost": {
            "type": "number",
            "example": 1200
          },
          "location": {
            "type": "string",
            "description": "Display / billing location hint until a dedicated column exists."
          },
          "requiresPayment": {
            "type": "boolean",
            "example": true,
            "description": "Force payment intent creation at registration for this tour. Requires module.finance capability."
          },
          "paymentMode": {
            "type": "string",
            "enum": [
              "offline_receipt"
            ],
            "description": "Denali pilot: offline receipt only (no online gateway)."
          }
        }
      },
...
      "TripDetailsOverviewDto": {
        "type": "object",
        "properties": {
          "mainDestination": {
            "type": "string"
          },
          "destinationRegion": {
            "type": "string"
          },
          "tourThemeIds": {
            "description": "Workspace catalog tour theme ids (`workspace_tour_themes.id`).",
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "tourThemeLabels": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Optional snapshot of theme id → display label for detail views when an id is no longer in the workspace catalog."
          },
          "tripStyles": {
            "type": "array",
            "description": "Multi-select execution style (adventure, relaxed, luxury, budget, familyFriendly, photography). Orthogonal to root `tourType` (mountain, city, …). Replaces a legacy singular overview style field (migrated to this array on read).",
            "items": {
              "type": "string",
              "enum": [
                "adventure",
                "relaxed",
                "luxury",
                "budget",
                "familyFriendly",
                "photography"
              ]
            }
          },
          "difficultyLevel": {
            "type": "number",
            "minimum": 1,
            "maximum": 10,
            "multipleOf": 0.5,
            "example": 4.5,
            "description": "Numeric difficulty rating on a 1–10 scale with 0.5 step granularity (allowed values: 1, 1.5, 2, …, 9.5, 10). Replaces the legacy enum."
          },
          "elevationGainMeters": {
            "type": "number"
          },
          "maxAltitudeMeters": {
            "type": "number"
          },
          "bestFor": {
            "deprecated": true,
            "description": "Deprecated. Use `participation.suitableFor` / `participation.notSuitableFor` instead. Retained for backward-compatible reads of older JSONB documents.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "shortIntro": {
            "type": "string",
            "maxLength": 250,
            "description": "Very short teaser (≤250 chars) for **tour cards / list previews / meta description**. The root tour `description` field holds the long-form story for the **full tour page**. If empty, clients should fall back to `description` when rendering cards."
          },
          "settingsRegionId": {
            "type": "string",
            "format": "uuid",
            "description": "Settings → Locations region id (wizard JSONB linkage)."
          },
          "settingsMainDestinationId": {
            "type": "string",
            "format": "uuid",
            "description": "Settings → Locations primary destination id (wizard)."
          },
          "secondaryDestinationIdsRaw": {
            "type": "string",
            "description": "Free-text / comma-separated secondary destination ids from the wizard when not normalized to UUID array."
          },
          "leaderUserIds": {
            "description": "Workspace tour leader ids.",
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "localGuideName": {
            "type": "string",
            "description": "Display name for the local guide (when not a workspace user)."
          },
          "startPoint": {
            "$ref": "#/components/schemas/TripDetailsLocationDataDto"
          },
          "summitPoint": {
            "$ref": "#/components/schemas/TripDetailsLocationDataDto"
          },
          "campPoint": {
            "$ref": "#/components/schemas/TripDetailsLocationDataDto"
          },
          "endPoint": {
            "$ref": "#/components/schemas/TripDetailsLocationDataDto"
          },
          "denaliTourKind": {
            "type": "string",
            "enum": [
              "mountain_day",
              "mountain_multi",
              "nature_day",
              "nature_multi",
              "desert_day",
              "desert_multi",
              "event_reading",
              "event_cinema"
            ],
            "description": "Denali 6-tab wizard tour kind slug (persisted for edit/clone round-trip)."
          },
          "nonAttendanceDetails": {
            "type": "string",
            "description": "Optional admin note for non-attendance (Denali wizard)."
          },
          "customServiceLabels": {
            "description": "Workspace-defined custom service labels (Denali wizard; also accepted on root create/update DTO).",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "TripDetailsItineraryDto": {
        "type": "object",
        "properties": {
          "highlights": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "includedVisits": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "excludedVisits": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "optionalActivities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "outline": {
            "type": "string",
            "description": "Free-text itinerary / program summary."
          },
          "programNotes": {
            "type": "string"
          },
          "specialExperiences": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dayPlans": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TripDetailsDayPlanDto"
            }
          },
          "segmentActivities": {
            "description": "Structured per-day segments from the tour wizard (distinct from legacy `dayPlans` rows).",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TripDetailsSegmentActivityDayDto"
            }
          }
        }
      },
...
      "TripDetailsDenaliTransportDto": {
        "type": "object",
        "properties": {
          "transportCost": {
            "type": "number",
            "description": "Organizer transport fee per person (Toman) — not دنگ."
          },
          "allowPersonalCar": {
            "type": "boolean",
            "description": "Bus/minibus/train: participants may use a personal car."
          },
          "dongAmount": {
            "type": "number",
            "description": "Personal car fuel share (دنگ) when allowed."
          }
        }
      },
      "TourTripDetailsDto": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "number",
            "example": 1,
            "minimum": 1
          },
          "overview": {
            "$ref": "#/components/schemas/TripDetailsOverviewDto"
          },
          "itinerary": {
            "$ref": "#/components/schemas/TripDetailsItineraryDto"
          },
          "participation": {
            "$ref": "#/components/schemas/TripDetailsParticipationDto"
          },
          "logistics": {
            "$ref": "#/components/schemas/TripDetailsLogisticsDto"
          },
          "policies": {
            "$ref": "#/components/schemas/TripDetailsPoliciesDto"
          },
          "transport": {
            "$ref": "#/components/schemas/TripDetailsDenaliTransportDto"
          },
          "photos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TripDetailsPhotoDto"
            }
          }
        }
      },
...
      "TourResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "22222222-2222-4222-8222-222222222222"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-05-01T08:00:00.000Z",
            "description": "Tour row creation time (UTC)."
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-05-01T08:05:00.000Z",
            "description": "Tour row last update time (UTC)."
          },
          "title": {
            "type": "string",
            "example": "Spring Camp 2026"
          },
          "description": {
            "type": "string",
            "example": "A two-day nature tour for members.",
            "nullable": true,
            "description": "Optional long-form description."
          },
          "totalCapacity": {
            "type": "number",
            "example": 30,
            "description": "Maximum capacity for the tour."
          },
          "acceptedCount": {
            "type": "number",
            "example": 10,
            "description": "Number of accepted registrations (capacity consumption)."
          },
          "lifecycleStatus": {
            "type": "string",
            "enum": [
              "DRAFT",
              "OPEN",
              "CLOSED",
              "CANCELLED"
            ],
            "example": "OPEN",
            "description": "Tour lifecycle state."
          },
          "chatLink": {
            "type": "string",
            "nullable": true,
            "example": "https://t.me/joinchat/example",
            "description": "Leader-managed communication link (e.g. Telegram)."
          },
          "costContext": {
            "type": "object",
            "example": {
              "currency": "USD",
              "totalCost": 1200
            },
            "nullable": true,
            "additionalProperties": true,
            "description": "Opaque pricing / operational JSON (JSONB)."
          },
          "autoAcceptRegistrations": {
            "type": "boolean",
            "nullable": true,
            "example": true,
            "description": "When false, registrations are not auto-accepted by tour policy."
          },
          "tourType": {
            "type": "string",
            "enum": [
              "mountain",
              "city",
              "desert",
              "nature",
              "cultural"
            ],
            "nullable": true,
            "example": "mountain",
            "description": "Top-level **category** (mountain, city, desert, nature, cultural). Execution-style selections live in `tripDetails.overview.tripStyles`."
          },
          "transportModes": {
            "type": "array",
            "example": [
              "bus",
              "train"
            ],
            "description": "Organized transport modes for this tour (multi-select). Empty when unset. No `mixed` slug.",
            "items": {
              "type": "string",
              "enum": [
                "bus",
                "train",
                "plane",
                "private_car"
              ]
            }
          },
          "formProfileSnapshot": {
            "type": "string",
            "nullable": true,
            "description": "Resolved form profile at creation (`general`, `urban_event`, …); optional frozen snapshot on the row."
          },
          "destinationId": {
            "type": "object",
            "format": "uuid",
            "nullable": true,
            "description": "FK to workspace_destinations when the tour is linked to a Settings destination."
          },
          "destinationName": {
            "type": "string",
            "nullable": true,
            "example": "Damavand",
            "description": "Destination display name (from Settings) when `destinationId` is set."
          },
          "destinationRegionName": {
            "type": "string",
            "nullable": true,
            "example": "Mazandaran",
            "description": "Region name for the linked destination, for list/detail UI."
          },
          "details": {
            "type": "object",
            "nullable": true,
            "properties": {
              "destinationName": {
                "type": "string",
                "nullable": true,
                "example": "Damavand"
              },
              "elevationM": {
                "type": "number",
                "nullable": true,
                "example": 5671
              },
              "difficulty": {
                "enum": [
                  "easy",
                  "moderate",
                  "hard",
                  "technical"
                ],
                "nullable": true,
                "example": "hard"
              },
              "durationDays": {
                "type": "number",
                "nullable": true,
                "example": 3
              },
              "meetingPoint": {
                "type": "string",
                "nullable": true,
                "example": "Azadi Square, Gate 3"
              },
              "itinerary": {
                "type": "array",
                "nullable": true,
                "items": {
                  "type": "object",
                  "properties": {
                    "day": {
                      "type": "number",
                      "example": 1
                    },
                    "title": {
                      "type": "string",
                      "nullable": true,
                      "example": "Approach"
                    },
                    "description": {
                      "type": "string",
                      "nullable": true,
                      "example": "Transfer to base camp."
                    },
                    "distanceKm": {
                      "type": "number",
                      "nullable": true,
                      "example": 8
                    },
                    "elevationGainM": {
                      "type": "number",
                      "nullable": true,
                      "example": 450
                    }
                  }
                }
              },
              "tripDetails": {
                "type": "object",
                "nullable": true,
                "additionalProperties": true,
                "description": "Structured trip details (JSONB). Separate from tour `description`."
              }
            }
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "title",
          "totalCapacity",
          "acceptedCount",
          "lifecycleStatus",
          "transportModes"
        ]
      },
      "UpdateTourDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Damavand summit — 2-day climb (updated)",
            "minLength": 10,
            "maxLength": 120
          },
          "total_capacity": {
            "type": "number",
            "example": 40,
            "minimum": 0
          },
          "lifecycle_status": {
            "type": "string",
            "example": "OPEN",
            "enum": [
              "DRAFT",
              "OPEN",
              "CLOSED",
              "CANCELLED"
            ]
          },
          "description": {
            "type": "string",
            "example": "Updated description",
            "description": "Long-form copy for the **tour detail page**. When publishing or opening a tour to the public, this should be populated for consistency with list/detail UX. Teaser text for cards belongs in `tripDetails.overview.shortIntro`, not here."
          },
          "chat_link": {
            "type": "string",
            "example": "https://t.me/joinchat/updated"
          },
          "cost_context": {
            "example": {
              "currency": "USD",
              "totalCost": 1200
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/CostContextDto"
              }
            ]
          },
          "autoAcceptRegistrations": {
            "type": "boolean",
            "example": true
          },
          "tourType": {
            "type": "string",
            "enum": [
              "mountain",
              "city",
              "desert",
              "nature",
              "cultural"
            ],
            "example": "mountain",
            "description": "Top-level **category** of the tour. Use `tripDetails.overview.tripStyles` for the execution-style selections (adventure, luxury, …)."
          },
          "formProfile": {
            "type": "string",
            "enum": [
              "general",
              "mountain_outdoor",
              "nature_trip",
              "urban_event",
              "cinema_event",
              "cultural_tour",
              "denali_pilot"
            ],
            "example": "urban_event",
            "description": "Optional client hint (ignored on update). The authoritative form profile is resolved server-side from `workspace_tour_wizard_templates.base_profile` for strip, invariants, and `formProfileSnapshot`."
          },
          "transportModes": {
            "type": "array",
            "example": [
              "plane"
            ],
            "description": "Replace tour transport modes (multi-select). Send `[]` or `null` to clear. No `mixed` — combine concrete modes.",
            "items": {
              "type": "string",
              "enum": [
                "bus",
                "train",
                "plane",
                "private_car"
              ]
            }
          },
          "destinationId": {
            "type": "object",
            "format": "uuid",
            "nullable": true,
            "description": "Workspace destination from Settings → Locations. Send `null` to clear."
          },
          "destinationName": {
            "type": "string",
            "example": "Damavand"
          },
          "elevationM": {
            "type": "number",
            "example": 5671
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "easy",
              "moderate",
              "hard",
              "technical"
            ],
            "example": "moderate"
          },
          "durationDays": {
            "type": "number",
            "example": 3,
            "minimum": 1,
            "maximum": 60,
            "description": "Inclusive day count. Re-derived server-side from `tripDetails.logistics.departureDate`/`returnDate` when those are part of the patch."
          },
          "meetingPoint": {
            "type": "string",
            "example": "Azadi Square, Gate 3"
          },
          "itinerary": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "day": {
                  "type": "number",
                  "example": 1
                },
                "title": {
                  "type": "string",
                  "example": "Base camp approach"
                },
                "description": {
                  "type": "string",
                  "example": "Drive and acclimatization hike."
                },
                "distanceKm": {
                  "type": "number",
                  "example": 8
                },
                "elevationGainM": {
                  "type": "number",
                  "example": 450
                }
              }
            }
          },
          "tripDetails": {
            "description": "Structured trip details (JSON). Patches are deep-merged into the stored document (omit a key to leave it unchanged; arrays replace when sent). Pass root `null` to clear the whole blob. Separate from `description` marketing text.",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/TourTripDetailsDto"
              }
            ]
          },
          "customServiceLabels": {
            "example": [
              "صبحانه",
              "نیسان"
            ],
            "description": "Workspace-defined custom service labels from the Denali wizard (persisted under tripDetails.overview.customServiceLabels).",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "WorkspaceEquipmentItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "category",
          "description",
          "icon",
          "isActive",
          "sortOrder",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateEquipmentItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1
          },
          "slug": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1
          },
          "category": {
            "type": "object",
            "nullable": true,
            "maxLength": 80
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "icon": {
            "type": "object",
            "nullable": true,
            "maxLength": 120
          },
          "isActive": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "number"
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
...
      "WorkspaceTourThemeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "slug": {
            "type": "string",
            "maxLength": 120
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "number"
          },
          "formProfile": {
            "type": "string",
            "enum": [
              "general",
              "mountain_outdoor",
              "nature_trip",
              "urban_event",
              "cinema_event",
              "cultural_tour",
              "denali_pilot"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "description",
          "isActive",
          "sortOrder",
          "formProfile",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateWorkspaceTourThemeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1
          },
          "slug": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "number"
          },
          "formProfile": {
            "type": "string",
            "enum": [
              "general",
              "mountain_outdoor",
              "nature_trip",
              "urban_event",
              "cinema_event",
              "cultural_tour",
              "denali_pilot"
            ],
            "description": "Tour creation wizard profile bound to this theme.",
            "default": "general"
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
...
    }
  }
}